+++ src/main/java/com/clubber/ClubberServer/domain/admin/service/AdminAuthService.java
@@ -3,8 +3,8 @@ package com.clubber.ClubberServer.domain.admin.service;
 import com.clubber.ClubberServer.domain.admin.domain.Admin;
 import com.clubber.ClubberServer.domain.admin.dto.CreateAdminsLoginRequest;
 import com.clubber.ClubberServer.domain.admin.dto.CreateAdminsLoginResponse;
-import com.clubber.ClubberServer.domain.admin.impl.TokenAppender;
-import com.clubber.ClubberServer.domain.admin.impl.TokenReader;
+import com.clubber.ClubberServer.domain.admin.implement.TokenAppender;
+import com.clubber.ClubberServer.domain.admin.implement.TokenReader;
 import com.clubber.ClubberServer.domain.admin.validator.AdminValidator;
 import com.clubber.ClubberServer.domain.auth.vo.TokenVO;
 import com.clubber.ClubberServer.global.config.security.SecurityUtils;
